Add registers to initiate a beacon transmission.
Writing to the TXLEN register will cause the badge to initiate a beacon
transmission for 1s using the contents of the TXMAGIC and TXDATA registers.
The beacon will be sent at the fast advertising rate for 1s after which
the badge will return to the normal beacon at the slow rate.

There is some concern over the maximum beacon payload, so we have also
reduced the size of the badge name string to guarantee space for a few
bytes of payload.