Use a periodic timer to clear magic beacon effects.
It has rarely been observed where badges will fail to clear an emote
or awoo status, I suspect that this is being caused by a race between
the beacon refresh worker and the reception of a new beacon. As an
attempt to make the badge more reliable, we move the expiration
timing into a periodic timer so that it gets caught on the next
iteration.